Extract radial-gradient frame source
